+++ CesiumGltf/include/CesiumGltf/PropertyTexturePropertyView.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "CesiumGltf/ImageCesium.h"
+#include "CesiumGltf/KhrTextureTransform.h"
 #include "CesiumGltf/PropertyTextureProperty.h"
 #include "CesiumGltf/PropertyTransformations.h"
 #include "CesiumGltf/PropertyTypeTraits.h"
@@ -12,6 +13,7 @@
 #include <cassert>
 #include <cmath>
 #include <cstdint>
+#include <optional>
 
 namespace CesiumGltf {
 /**
@@ -197,6 +199,8 @@ std::array<uint8_t, 4> sampleNearestPixel(
     const double u,
     const double v);
 
+#pragma region Non - normalized property
+
 /**
  * @brief A view of the data specified by a {@link PropertyTextureProperty}.
  *
@@ -235,7 +239,9 @@ public:
         _pImage(nullptr),
         _texCoordSetIndex(0),
         _channels(),
-        _swizzle() {}
+        _swizzle(),
+        _applyTextureTransform(false),
+        _textureTransform(std::nullopt) {}
 
   /**
    * @brief Constructs an invalid instance for an erroneous property.
@@ -248,7 +254,9 @@ public:
         _pImage(nullptr),
         _texCoordSetIndex(0),
         _channels(),
-        _swizzle() {
+        _swizzle(),
+        _applyTextureTransform(false),
+        _textureTransform(std::nullopt) {
     assert(
         this->_status != PropertyTexturePropertyViewStatus::Valid &&
         "An empty property view should not be constructed with a valid status");
@@ -268,7 +276,9 @@ public:
         _pImage(nullptr),
         _texCoordSetIndex(0),
         _channels(),
-        _swizzle() {
+        _swizzle(),
+        _applyTextureTransform(false),
+        _textureTransform(std::nullopt) {
     if (this->_status != PropertyTexturePropertyViewStatus::Valid) {
       // Don't override the status / size if something is wrong with the class
       // property's definition.
@@ -290,23 +300,43 @@ public:
   /**
    * @brief Construct a view of the data specified by a {@link PropertyTextureProperty}.
    *
+   * A property texture property may contain the `KHR_texture_transform`
+   * extension, which transforms the texture coordinates used to sample the
+   * texture. The extension may also override the TEXCOORD set index that was
+   * originally specified by the property texture property.
+   *
+   * If a view is constructed with applyKhrTextureTransformExtension set to
+   * true, the view will automatically apply the texture transform to any UV
+   * coordinates used to sample the texture. If the extension defines its own
+   * TEXCOORD set index, it will override the original value.
+   *
+   * Otherwise, if the flag is set to false, UVs will not be transformed and
+   * the original TEXCOORD set index will be preserved. The extension's values
+   * may still be retrieved using getTextureTransform, if desired.
+   *
    * @param property The {@link PropertyTextureProperty}
    * @param classProperty The {@link ClassProperty} this property conforms to.
    * @param sampler The {@link Sampler} used by the property.
    * @param image The {@link ImageCesium} used by the property.
-   * @param channels The code from {@link PropertyTextureProperty::channels}.
+   * @param channels The value of {@link PropertyTextureProperty::channels}.
+   * @param applyKhrTextureTransformExtension Whether to automatically apply the
+   * `KHR_texture_transform` extension to the property texture property, if it
+   * exists.
    */
   PropertyTexturePropertyView(
       const PropertyTextureProperty& property,
       const ClassProperty& classProperty,
       const Sampler& sampler,
-      const ImageCesium& image) noexcept
+      const ImageCesium& image,
+      bool applyKhrTextureTransformExtension = false) noexcept
       : PropertyView<ElementType, false>(classProperty, property),
         _pSampler(&sampler),
         _pImage(&image),
         _texCoordSetIndex(property.texCoord),
         _channels(property.channels),
-        _swizzle() {
+        _swizzle(),
+        _applyTextureTransform(applyKhrTextureTransformExtension),
+        _textureTransform(std::nullopt) {
     if (this->_status != PropertyTexturePropertyViewStatus::Valid) {
       return;
     }
@@ -330,6 +360,13 @@ public:
       default:
         assert(false && "A valid channels vector must be passed to the view.");
       }
+    }
+
+    const ExtensionKhrTextureTransform* pTextureTransform =
+        property.getExtension<ExtensionKhrTextureTransform>();
+
+    if (pTextureTransform) {
+      this->_textureTransform = KhrTextureTransform(*pTextureTransform);
     }
   }
 
@@ -392,19 +429,35 @@ public:
         this->_status == PropertyTexturePropertyViewStatus::Valid &&
         "Check the status() first to make sure view is valid");
 
-    double wrappedU = applySamplerWrapS(u, this->_pSampler->wrapS);
-    double wrappedV = applySamplerWrapT(v, this->_pSampler->wrapT);
+    if (this->_applyTextureTransform && this->_textureTransform) {
+      glm::dvec2 transformedUvs = this->_textureTransform->applyTransform(u, v);
+      u = transformedUvs.x;
+      v = transformedUvs.y;
+    }
+
+    u = applySamplerWrapS(u, this->_pSampler->wrapS);
+    v = applySamplerWrapT(v, this->_pSampler->wrapT);
 
     std::array<uint8_t, 4> sample =
-        sampleNearestPixel(*this->_pImage, this->_channels, wrappedU, wrappedV);
+        sampleNearestPixel(*this->_pImage, this->_channels, u, v);
     return assembleValueFromChannels<ElementType>(
         gsl::span(sample.data(), this->_channels.size()));
   }
 
   /**
    * @brief Get the texture coordinate set index for this property.
+   *
+   * If applyKhrTextureTransformExtension is true, and if the property texture
+   * property contains the `KHR_texture_transform` extension, this will return
+   * the value from the extension, as it is meant to override the original
+   * index. However, if the extension does not specify a TEXCOORD set index,
+   * then the original index of the property texture property is returned.
    */
   int64_t getTexCoordSetIndex() const noexcept {
+    if (this->_applyTextureTransform && this->_textureTransform) {
+      return this->_textureTransform->getTexCoordSetIndex().value_or(
+          this->_texCoordSetIndex);
+    }
     return this->_texCoordSetIndex;
   }
 
@@ -437,13 +490,36 @@ public:
    */
   const std::string& getSwizzle() const noexcept { return this->_swizzle; }
 
+  /**
+   * @brief Get the KHR_texture_transform for this property texture property, if
+   * it exists.
+   *
+   * Even if this view was constructed with applyKhrTextureTransformExtension
+   * set to false, it will save the extension's values, and they may be
+   * retrieved through this function.
+   *
+   * If this view was constructed with applyKhrTextureTransformExtension set to
+   * true, any texture coordinates passed into `get` or `getRaw` will be
+   * automatically transformed, so there's no need to re-apply the transform
+   * here.
+   */
+  std::optional<KhrTextureTransform> getTextureTransform() const noexcept {
+    return this->_textureTransform;
+  }
+
 private:
   const Sampler* _pSampler;
   const ImageCesium* _pImage;
   int64_t _texCoordSetIndex;
   std::vector<int64_t> _channels;
   std::string _swizzle;
+  bool _applyTextureTransform;
+  std::optional<KhrTextureTransform> _textureTransform;
 };
+
+#pragma endregion
+
+#pragma region Normalized property
 
 /**
  * @brief A view of the normalized data specified by a
@@ -468,7 +544,9 @@ public:
         _pImage(nullptr),
         _texCoordSetIndex(0),
         _channels(),
-        _swizzle() {}
+        _swizzle(),
+        _applyTextureTransform(false),
+        _textureTransform(std::nullopt) {}
 
   /**
    * @brief Constructs an invalid instance for an erroneous property.
@@ -481,7 +559,9 @@ public:
         _pImage(nullptr),
         _texCoordSetIndex(0),
         _channels(),
-        _swizzle() {
+        _swizzle(),
+        _applyTextureTransform(false),
+        _textureTransform(std::nullopt) {
     assert(
         this->_status != PropertyTexturePropertyViewStatus::Valid &&
         "An empty property view should not be constructed with a valid status");
@@ -501,7 +581,9 @@ public:
         _pImage(nullptr),
         _texCoordSetIndex(0),
         _channels(),
-        _swizzle() {
+        _swizzle(),
+        _applyTextureTransform(false),
+        _textureTransform(std::nullopt) {
     if (this->_status != PropertyTexturePropertyViewStatus::Valid) {
       // Don't override the status / size if something is wrong with the class
       // property's definition.
@@ -523,23 +605,43 @@ public:
   /**
    * @brief Construct a view of the data specified by a {@link PropertyTextureProperty}.
    *
+   * A property texture property may contain the `KHR_texture_transform`
+   * extension, which transforms the texture coordinates used to sample the
+   * texture. The extension may also override the TEXCOORD set index that was
+   * originally specified by the property texture property.
+   *
+   * If a view is constructed with applyKhrTextureTransformExtension set to
+   * true, the view will automatically apply the texture transform to any UV
+   * coordinates used to sample the texture. If the extension defines its own
+   * TEXCOORD set index, it will override the original value.
+   *
+   * Otherwise, if the flag is set to false, UVs will not be transformed and
+   * the original TEXCOORD set index will be preserved. The extension's values
+   * may still be retrieved using getTextureTransform, if desired.
+   *
    * @param property The {@link PropertyTextureProperty}
    * @param classProperty The {@link ClassProperty} this property conforms to.
    * @param sampler The {@link Sampler} used by the property.
    * @param image The {@link ImageCesium} used by the property.
    * @param channels The value of {@link PropertyTextureProperty::channels}.
+   * @param applyKhrTextureTransformExtension Whether to automatically apply the
+   * `KHR_texture_transform` extension to the property texture property, if it
+   * exists.
    */
   PropertyTexturePropertyView(
       const PropertyTextureProperty& property,
       const ClassProperty& classProperty,
       const Sampler& sampler,
-      const ImageCesium& image) noexcept
+      const ImageCesium& image,
+      bool applyKhrTextureTransformExtension = false) noexcept
       : PropertyView<ElementType, true>(classProperty, property),
         _pSampler(&sampler),
         _pImage(&image),
         _texCoordSetIndex(property.texCoord),
         _channels(property.channels),
-        _swizzle() {
+        _swizzle(),
+        _applyTextureTransform(applyKhrTextureTransformExtension),
+        _textureTransform(std::nullopt) {
     if (this->_status != PropertyTexturePropertyViewStatus::Valid) {
       return;
     }
@@ -562,6 +664,13 @@ public:
       default:
         assert(false && "A valid channels vector must be passed to the view.");
       }
+    }
+
+    const ExtensionKhrTextureTransform* pTextureTransform =
+        property.getExtension<ExtensionKhrTextureTransform>();
+
+    if (pTextureTransform) {
+      this->_textureTransform = KhrTextureTransform(*pTextureTransform);
     }
   }
 
@@ -651,11 +760,17 @@ public:
         this->_status == PropertyTexturePropertyViewStatus::Valid &&
         "Check the status() first to make sure view is valid");
 
-    double wrappedU = applySamplerWrapS(u, this->_pSampler->wrapS);
-    double wrappedV = applySamplerWrapT(v, this->_pSampler->wrapT);
+    if (this->_applyTextureTransform && this->_textureTransform) {
+      glm::dvec2 transformedUvs = this->_textureTransform->applyTransform(u, v);
+      u = transformedUvs.x;
+      v = transformedUvs.y;
+    }
+
+    u = applySamplerWrapS(u, this->_pSampler->wrapS);
+    v = applySamplerWrapT(v, this->_pSampler->wrapT);
 
     std::array<uint8_t, 4> sample =
-        sampleNearestPixel(*this->_pImage, this->_channels, wrappedU, wrappedV);
+        sampleNearestPixel(*this->_pImage, this->_channels, u, v);
 
     return assembleValueFromChannels<ElementType>(
         gsl::span(sample.data(), this->_channels.size()));
@@ -663,8 +778,18 @@ public:
 
   /**
    * @brief Get the texture coordinate set index for this property.
+   *
+   * If applyKhrTextureTransformExtension is true, and if the property texture
+   * property contains the `KHR_texture_transform` extension, this will return
+   * the value from the extension, as it is meant to override the original
+   * index. However, if the extension does not specify a TEXCOORD set index,
+   * then the original index of the property texture property is returned.
    */
   int64_t getTexCoordSetIndex() const noexcept {
+    if (this->_applyTextureTransform && this->_textureTransform) {
+      return this->_textureTransform->getTexCoordSetIndex().value_or(
+          this->_texCoordSetIndex);
+    }
     return this->_texCoordSetIndex;
   }
 
@@ -697,12 +822,32 @@ public:
    */
   const std::string& getSwizzle() const noexcept { return this->_swizzle; }
 
+  /**
+   * @brief Get the KHR_texture_transform for this property texture property, if
+   * it exists.
+   *
+   * Even if this view was constructed with applyKhrTextureTransformExtension
+   * set to false, it will save the extension's values, and they may be
+   * retrieved through this function.
+   *
+   * If this view was constructed with applyKhrTextureTransformExtension set to
+   * true, any texture coordinates passed into `get` or `getRaw` will be
+   * automatically transformed, so there's no need to re-apply the transform
+   * here.
+   */
+  std::optional<KhrTextureTransform> getTextureTransform() const noexcept {
+    return this->_textureTransform;
+  }
+
 private:
   const Sampler* _pSampler;
   const ImageCesium* _pImage;
   int64_t _texCoordSetIndex;
   std::vector<int64_t> _channels;
   std::string _swizzle;
+  bool _applyTextureTransform;
+  std::optional<KhrTextureTransform> _textureTransform;
 };
+#pragma endregion
 
 } // namespace CesiumGltf